+++ DocumentModel/Classes/EditBox.cs
@@ -33,7 +33,7 @@ public partial class EditBox
   /// <summary>
   /// maxLength, this property is only available in Office 2010 and later.
   /// </summary>
-  public Int32? MaxLength { get; set; }
+  public Int64? MaxLength { get; set; }
   
   /// <summary>
   /// getText, this property is only available in Office 2010 and later.
