+++ DocumentModel/Math/Classes/ArgumentProperties.cs
@@ -8,6 +8,6 @@ public partial class ArgumentProperties
   /// <summary>
   /// Argument Size.
   /// </summary>
-  public Int32? ArgumentSize { get; set; }
+  public Int64? ArgumentSize { get; set; }
   
 }
