+++ DocumentModel/Math/Classes/Break.cs
@@ -8,11 +8,11 @@ public partial class Break
   /// <summary>
   /// Index of Operator to Align To
   /// </summary>
-  public Int32? AlignAt { get; set; }
+  public Int64? AlignAt { get; set; }
   
   /// <summary>
   /// Index of Operator to Align To
   /// </summary>
-  public Int32? Val { get; set; }
+  public Int64? Val { get; set; }
   
 }
