+++ DocumentModel/UI/Classes/EditBox.cs
@@ -33,7 +33,7 @@ public partial class EditBox
   /// <summary>
   /// maxLength
   /// </summary>
-  public Int32? MaxLength { get; set; }
+  public Int64? MaxLength { get; set; }
   
   /// <summary>
   /// getText
