+++ DocumentModel/Wordprocessing/Classes/FootnoteEndnoteReferenceType.cs
@@ -13,6 +13,6 @@ public partial class FootnoteEndnoteReferenceType
   /// <summary>
   /// Footnote/Endnote ID Reference
   /// </summary>
-  public Int32? Id { get; set; }
+  public Int64? Id { get; set; }
   
 }
