+++ DocumentModel/Wordprocessing/Classes/NumberingSymbolRunProperties.cs
@@ -103,7 +103,7 @@ public partial class NumberingSymbolRunProperties
   /// <summary>
   /// CharacterScale.
   /// </summary>
-  public Int32? CharacterScale { get; set; }
+  public Int64? CharacterScale { get; set; }
   
   /// <summary>
   /// Kern.
