+++ DocumentModel/Wordprocessing/Classes/RunPropertiesBaseStyle.cs
@@ -103,7 +103,7 @@ public partial class RunPropertiesBaseStyle
   /// <summary>
   /// CharacterScale.
   /// </summary>
-  public Int32? CharacterScale { get; set; }
+  public Int64? CharacterScale { get; set; }
   
   /// <summary>
   /// Kern.
