+++ DocumentModel/Wordprocessing/Classes/StyleRunProperties.cs
@@ -103,7 +103,7 @@ public partial class StyleRunProperties
   /// <summary>
   /// CharacterScale.
   /// </summary>
-  public Int32? CharacterScale { get; set; }
+  public Int64? CharacterScale { get; set; }
   
   /// <summary>
   /// Kern.
